BOTServiceCenterPH/requirements.txt
Flask==2.3.3
scikit-learn==1.7.1
spacy==3.8.7
sentence-transformers==2.7.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.67.0
torch>=1.10.0
transformers>=4.25.0
Unidecode
colorama
pystray
Pillow
python-dotenv
